Split current tests for `update_blog_status()`
The current tests for `update_blog_status()` mirrored the tests for `update_blog_details()` in many ways and can be split in the same way. A noticeable difference is that the the matching actions fire even when no change is made to a field.

See #30080

Built from https://develop.svn.wordpress.org/trunk@30785