Posts: Remove the slug from Quick Edit for posts that are not `is_post_type_viewable()`.
This is a followup to [43728], which added the same check, but using the `publicly_queryable` option. This behaviour is incorrect for `_builtin` post types, which go by the logic in `is_post_type_viewable()`, instead.

Merges [43855] from the 5.0 branch to trunk.

Props pento.
See #43278.

Built from https://develop.svn.wordpress.org/trunk@44221