Editor: Use explicit context instead of a trailing space for Edit string in TinyMCE.
In [42523], `trunk` was updated to use explicit context instead of a trailing space for the Edit string in TinyMCE plugins. [43472] later backported this change, but the string in `class-wp-editor.php` was missed. This fixes the `Edit` string so not display as `Edit|button` in the 5.0 branch.

Props dimadin.

See #42543, #44134.
Fixes #44919.
Built from https://develop.svn.wordpress.org/branches/5.0@44401