Theme Editor: Remove unused `$has_templates` boolean.
The variable is a remnant from when the Theme Editor still had the "Templates" heading and is unused in the current logic.

The condition it was a part of would always evaluate to true for a child theme, regardless of the `$has_templates` value.

Props dboy1988.
Fixes #50199.
Built from https://develop.svn.wordpress.org/trunk@47844