Build/Test Tools: Allow the desired version of PHPUnit to be passed to the local Docker environment.
This introduces the `LOCAL_PHPUNIT` environment variable, which allows the desired version of PHPUnit to be specified when running the PHP tests within the local Docker environment.

Because support for newer versions of PHPUnit is not backported, some versions of PHP need the ability to run multiple versions of PHPUnit for different branches. This adds the flexibility needed to use the Docker environment within those older branches to run the PHP tests.

Props johnbillion, SergeyBiryukov.
Merges [49362] to the 5.4 branch.
See #50042.
Built from https://develop.svn.wordpress.org/branches/5.4@49364